pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[project]
name = "eegpp2-beta"
version = "0.1.4"
authors = [
    { name = "Vũ Anh Tuấn", email = "vuanhtuan1407@gmail.com" },
]
description = "EEG Phrase Predictor ver 2"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    'dropbox>=12.0.2',
    'joblib>=1.4.2',
    'numpy>=2.1.1',
    'tqdm>=4.66.5',
    'torchinfo>=1.8.0',
    'scipy>=1.14.1',
    'pyyaml>=6.0.2',
    'pywavelets>=1.7.0',
    'ptwt>=0.1.9',
    'pandas>=2.2.2',
    'lightning>=2.4.0',
    'scikit-learn>=1.5.2',
    'seaborn>=0.13.2',
]
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
eegpp2 = ["configs/*.yml"]
